fix(frontend): move SubmittedMessagesList to relay-workflows-lib
