+++ frontend/relay-workflows-lib/lib/main.ts
@@ -1,7 +1,7 @@
 export { default as Workflow } from "./components/WorkflowRelay";
 export { default as Submission } from "./components/SubmissionForm";
 export { default as RetriggerWorkflow } from "./query-components/RetriggerWorkflow";
-export { default as SubmittedMessagesList } from "workflows-lib/lib/components/workflow/SubmittedMessagesList";
+export { default as SubmittedMessagesList } from "./components/SubmittedMessagesList";
 export { default as WorkflowListFilterDrawer } from "./components/WorkflowListFilterDrawer";
 export { default as WorkflowsListView } from "./views/WorkflowsListView";
 export { default as RepositoryLink } from "./query-components/RepositoryLink";
